🚀 Feat : API 호출 기반으로 랜더링 작업중
