+++ src/pages/PostDetail/LikeButton/LikeButton.css
@@ -1,7 +1,6 @@
 .like-btn-container {
   display: flex;
   width: 4.8rem;
-  margin-top: 2rem;
   margin-inline: auto;
   border-radius: 60px;
   align-items: center;
